+++ src/test/java/roomescape/controller/ReservationApiControllerTest.java
@@ -1,0 +1,161 @@
+package roomescape.controller;
+
+import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.stream.Stream;
+
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.is;
+
+class ReservationApiControllerTest {
+
+    @DisplayName("Reservation 목록을 가져온다.")
+    @Test
+    void reservationAdminPageTest() {
+        RestAssured.given().log().all()
+                .when().get("/admin/reservation")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @DisplayName("Reservation 목록 내용 갯수를 검사한다")
+    @Test
+    void reservationTest() {
+        RestAssured.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(0));
+    }
+
+    @DisplayName("Reservation 입력 테스트")
+    @Test
+    void addReservationTest() {
+        Map<String, String> params = new HashMap<>();
+        params.put("name", "브라운");
+        params.put("date", "2023-08-05");
+        params.put("time", "15:40");
+
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(params)
+                .when().post("/reservations")
+                .then().log().all()
+                .statusCode(200)
+                .body("id", is(1));
+
+        RestAssured.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(1));
+    }
+
+    @DisplayName("Reservation 입력 후 ResponseBody의 시간 포맷 검사")
+    @Test
+    void reservationResponseTest() {
+        Map<String, String> params = new HashMap<>();
+        params.put("name", "브라운");
+        params.put("date", "2023-08-05");
+        params.put("time", "15:40");
+
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(params)
+                .when().post("/reservations")
+                .then().log().all()
+                .statusCode(200);
+
+        RestAssured.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .body("[0].time", equalTo("15:40"));
+    }
+
+    @DisplayName("Group에 존재하는 Id만 삭제할 수 있다.")
+    @Test
+    void deleteReservationTest() {
+        Map<String, String> params = new HashMap<>();
+        params.put("name", "브라운");
+        params.put("date", "2023-08-05");
+        params.put("time", "15:40");
+
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(params)
+                .when().post("/reservations")
+                .then().log().all()
+                .statusCode(200);
+
+        RestAssured.given().log().all()
+                .when().delete("/reservations/1")
+                .then().log().all()
+                .statusCode(200);
+
+        RestAssured.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(0));
+    }
+
+    @DisplayName("Reservation 요청에 null 존재할 수 없다.")
+    @ParameterizedTest
+    @MethodSource("invalidReservationRequestArguments")
+    void exceptNullReservationTest(String name, String date, String time) {
+        Map<String, String> params = new HashMap<>();
+        params.put("name", name);
+        params.put("date", date);
+        params.put("time", time);
+
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(params)
+                .when().post("/reservations")
+                .then().log().all()
+                .statusCode(400);
+    }
+
+    static Stream<Arguments> invalidReservationRequestArguments() {
+        return Stream.of(
+                Arguments.of(null, String.valueOf(LocalDate.now()), String.valueOf(LocalTime.now())),
+                Arguments.of("가이온", null, String.valueOf(LocalTime.now())),
+                Arguments.of("가이온", String.valueOf(LocalDate.now()), null)
+        );
+    }
+
+    @DisplayName("존재하지 않는 Id의 Reservation을 삭제할 수 없다")
+    @Test
+    void invalidReservationIdTest() {
+        RestAssured.given().log().all()
+                .when().delete("/reservations/5")
+                .then().log().all()
+                .statusCode(404);
+    }
+
+    @DisplayName("올바른 시간의 포멧만 요청 가능하다.")
+    @Test
+    void invalidReservationTimeTest() {
+        Map<String, String> params = new HashMap<>();
+        params.put("name", "브라운");
+        params.put("date", "2023-08-05");
+        params.put("time", "15:40:00");
+
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(params)
+                .when().post("/reservations")
+                .then().log().all()
+                .statusCode(400);
+    }
+}
