+++ src/test/java/roomescape/controller/ReservationPageControllerTest.java
@@ -1,0 +1,17 @@
+package roomescape.controller;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class ReservationPageControllerTest {
+
+    @DisplayName("WelcomePage 테스트")
+    @Test
+    void welcomePageTest() {
+        RestAssured.given().log().all()
+                .when().get("/admin")
+                .then().log().all()
+                .statusCode(200);
+    }
+}
